(MoritzLangenstein) First attempt at build pipeline integration
