+++ .dockerignore
@@ -2,6 +2,6 @@
 
 !LICENSE
 !VERSION
-!fastentrypoints.py
 !setup.py
 !iaso
+!athena
