+++ MANIFEST.in
@@ -1,2 +1,1 @@
 include VERSION
-include fastentrypoints.py
